requirements.txt
django-organizations~=2.1.0
django-extensions~=3.2.2
django-allauth~=0.54.0
django-phonenumber-field~=6.1.0
phonenumbers~=8.13.0
immunity-utils[rest,celery] @ https://github.com/edge-servers/immunity-utils/tarball/master
packaging
django-sesame~=2.4.0
